git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@29 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/apps/mplayerc/PlayerSubresyncBar.cpp
@@ -1339,7 +1339,7 @@ bool CPlayerSubresyncBar::ShiftSubtitle(int nItem, long lValue, __int64& rtPos)
 		UpdatePreview();
 		SaveSubtitle();
 		bRet = true;
-		rtPos = m_subtimes[nItem].newstart * 10000;
+		rtPos = (__int64)m_subtimes[nItem].newstart * 10000;
 	}
 	return bRet;
 }
